requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
alembic
requests
python-dotenv
apscheduler
